@@ -4,46 +4,61 @@
   "description": "System F",
   "main": "index.js",
   "scripts": {
-    "test": "mocha tests/tests.js -w"
+    "start": "./node_modules/.bin/webpack-dev-server --inline --host 127.0.0.1",
+    "test": "./node_modules/karma/bin/karma start"
   },
-  "author": "Bruno Dias <and.coffee.code@gmail.com>",
+  "author": "Bruno Dias <dias.h.bruno@gmail.com>",
   "license": "MIT",
   "repository": {
     "type": "git",
     "url": "git://github.com/andCoffeeCode/asdf.git"
   },
   "devDependencies": {
-    "browserify": "^13.0.0",
-    "mocha": "^2.4.5",
-    "purecss": "^0.6.0",
-    "should": "^8.2.2",
-    "vinyl-source-stream": "^1.1.0",
-
-    "eslint": "1.10.3",
-    "eslint-loader": "1.2.0",
-    "eslint-plugin-react": "3.15.0",
-    "babel": "5.8.23",
-    "babel-eslint": "5.0.0-beta10",
-    "babel-loader": "5.1.4",
-    "babel-plugin-react-display-name": "2.0.0",
-    "babel-plugin-react-transform": "1.1.1",
-    "browser-sync": "2.11.1",
-    "rimraf": "2.5.0",
-
-    "node-sass": "3.4.2",
-    "file-loader": "0.8.5",
+    "babel-core": "^6.7.7",
+    "babel-loader": "^6.2.4",
+    "babel-plugin-syntax-export-extensions": "^6.13.0",
+    "babel-plugin-transform-object-rest-spread": "^6.23.0",
+    "babel-preset-es2015": "^6.6.0",
+    "babel-preset-stage-2": "^6.24.1",
     "css-loader": "0.23.1",
+    "eslint": "^3.19.0",
+    "eslint-loader": "^1.7.1",
+    "eslint-plugin-react": "^7.0.1",
+    "file-loader": "0.8.5",
+    "karma": "^1.7.0",
+    "karma-coverage": "^1.1.1",
+    "karma-mocha": "^1.3.0",
+    "karma-phantomjs-launcher": "^1.0.4",
+    "karma-spec-reporter": "^0.0.31",
+    "karma-webpack": "^2.0.3",
+    "mocha": "^3.4.2",
+    "node-sass": "3.4.2",
+    "purecss": "^0.6.0",
+    "redux-devtools": "^3.4.0",
+    "redux-devtools-dock-monitor": "^1.1.2",
+    "redux-devtools-log-monitor": "^1.3.0",
     "sass-loader": "3.1.2",
+    "should": "^8.2.2",
     "style-loader": "0.13.0",
     "url-loader": "0.5.7",
-    "watch": "0.17.1",
-    "webpack": "1.12.11",
-    "webpack-dev-middleware": "1.4.0",
-    "webpack-hot-middleware": "2.6.0"
+    "webpack": "2.6.1",
+    "webpack-dev-server": "^2.4.5"
   },
   "dependencies": {
-    "jquery": "^2.2.1",
+    "babel-preset-react": "^6.24.1",
+    "daggy": "0.0.1",
+    "data.task": "^3.1.1",
+    "draggable": "^4.0.3",
+    "eslint": "^3.19.0",
+    "fantasy-eithers": "0.0.2",
+    "fantasy-helpers": "0.0.1",
+    "fantasy-options": "0.0.1",
+    "fantasy-tuples": "0.0.1",
     "lodash": "^4.5.1",
-    "mathjax": "^2.6.1"
+    "mathjax": "^2.6.1",
+    "react": "^15.5.4",
+    "react-dom": "^15.5.4",
+    "react-redux": "^5.0.5",
+    "redux": "^3.6.0"
   }
 }
