@@ -1,35 +1,63 @@
 /*global require */
 
-var $ = require('jquery');
-var asdf = require("../src/renderers.js");
-var objs = require("../src/objects.js");
-var ev = require("../src/eval.js");
+import "../css/main.css";
+import React, { Component } from 'react';
+import ReactDOM from 'react-dom';
+import store, { Provider, DevTools, connect } from "./state";
+import GraphView from 'app/ui/graph';
+import { List, View } from "app/ui/list";
 
-var create_type = document.getElementById("create-type");
-var create_fn   = document.getElementById("create-fn");
-
-create_type.onclick = function(ev) {
-  // console.log("type", ev);
-};
-
-create_fn.onclick = function(ev) {
-  // console.log("fn", ev);
-};
-
-function build(data) {
-  var exp = asdf.render(data);
-  document.body.appendChild(exp);
+function createFunction(text) {
+  return {
+    type: "createNewFunction",
+    text
+  };
+}
+function cancelFunction(text) {
+  return {
+    type: "cancelNewFunction",
+    text
+  };
 }
 
-var promise1 = $.ajax("id.json");
-var promise2 = $.ajax("typeU.json");
+function _CreateFunctionView(props) {
+  return (
+    <div className="view-container">
+      <h2>Create new function:</h2>
+      <label htmlFor="label">Label:</label><input id="label" type="text" name="label" />
+      <label htmlFor="body">Expression:</label><input id="body" type="text" name="body" />
+      <button className="pure-button" onClick={props.cancel}>Cancel</button>
+      <button className="pure-button" onClick={props.create}>Create</button>
+    </div>
+  );
+}
 
-$.when(promise1, promise2).done(function(data1, data2) {
-  var as = [data1[0], data2[0]];
-  for (var a in as) {
-    var data = as[a];
-    build(data);
+const CreateFunctionView = connect(
+  state => ({}),
+  dispatch => ({
+    create: text => dispatch(createFunction(text)),
+    cancel: text => dispatch(cancelFunction())
+  })
+)(_CreateFunctionView);
+
+class App extends Component {
+  render() {
+    return (      
+      <div>
+        <DevTools />
+        <div className="pure-g">
+          <List collection={['a', 'b', 'c']} />
+          <View>
+            <GraphView  />
+          </View>
+        </div>
+      </div>
+    );
   }
+}
 
-  build(ev.eval_exp(data1[0], data2[0]));
-});
+ReactDOM.render((
+  <Provider store={store}>
+    <App />
+  </Provider>
+), document.getElementById("application"));
